Add HTTPS server support with Zephyr
Add the support for running HTTPS servers with Zephyr using Zephyr
Secure Sockets. Using Secure Sockets enables in particular to easily
supply the TLS certificate and private key since it is enough to add
them to TLS Credentials subsystem and to provide the corresponding tag
through the CivetWeb configuration.

Note that TLS support is not added for the HTTP client API.

Signed-off-by: Thomas Altenbach <taltenbach@witekio.com>